+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixed
 - an issue preventing the import and export of JSON mod sets if the JSON file was too large ([#39][gui39])
+- Extraction failures now display the name of the originating package (#42)
 
 [gui39]: https://github.com/memo33/sc4pac-gui/issues/39